requirements.txt
asgiref==3.7.2
boto3==1.28.51
botocore==1.31.51
cffi==1.15.1
cryptography==37.0.4
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.4.0
gunicorn==19.6.0
pycparser==2.21
pycryptodome==3.17
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
PyJWT==2.4.0
python-dateutil==2.8.2
requests==2.31.0
